chat - core provided chat submenus in components
